requirements.txt
protobuf==5.26.1
python-dotenv==1.0.1
streamlit==1.33.0
